+++ src/test/resources/logback.xml
@@ -21,14 +21,17 @@
         <projectPackage>com.company.package2</projectPackage>
 
         <metaData>
-            <tabName>logbackTab</tabName>
-            <key>logbackKey1</key>
-            <value>logbackValue1</value>
-        </metaData>
-        <metaData>
-            <tabName>logbackTab</tabName>
-            <key>logbackKey2</key>
-            <value>logbackValue2</value>
+            <tab>
+                <name>logbackTab</name>
+                <key>
+                    <name>logbackKey1</name>
+                    <value>logbackValue1</value>
+                </key>
+                <key>
+                    <name>logbackKey2</name>
+                    <value>logbackValue2</value>
+                </key>
+            </tab>
         </metaData>
 
         <sendThreads>true</sendThreads>
